+++ src/linear-algebra/blas/gemm.jl
@@ -20,8 +20,8 @@ let
     A = zeros(Float32,ni,nk)
     B = zeros(Float32,nk,nj)
 
-    alpha = 1.5
-    beta = 1.2
+    alpha = Float32(1.5)
+    beta = Float32(1.2)
     for i = 1:ni, j = 1:nj
         C[i,j] = ((i*j+1) % ni) / ni
     end
