+++ src/linear-algebra/blas/trmm.jl
@@ -15,7 +15,7 @@ let
     A = zeros(Float32,m,m)
     B = zeros(Float32,m,n)
 
-    alpha = 1.5
+    alpha = Float32(1.5)
     for i = 1:m
         for j = 1:(i-1)
             A[i,j] = ((i+j) % m)/m
